try to solve cyclic dependencies by merging all 3d object classes
