+++ docs/src/example_persistence.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python3
 from typing import Any
 
-from nicegui import ui
+from nicegui import app, ui
 
+import rosys
 from rosys import persistence
 
 
@@ -10,7 +11,11 @@ class Model(persistence.PersistentModule):
 
     def __init__(self) -> None:
         super().__init__()
-        self.value: float = 1.0
+        self.value: float = 5.0
+
+    def set_value(self, value: float) -> None:
+        self.value = value
+        self.request_backup()
 
     def restore(self, data: dict[str, Any]) -> None:
         self.value = data.get('value', 1.0)
@@ -21,7 +26,13 @@ class Model(persistence.PersistentModule):
         }
 
 
-model = Model()
-ui.slider(min=0, max=10.0, step=0.1).bind_value(model, 'value').props('label-always')
+def start():
+    model = Model()
+    model.set_value(3.0)
+    ui.slider(min=0, max=10.0, step=0.1, on_change=model.request_backup).bind_value(
+        model, 'value').props('label-always')
+
+
+app.on_startup(start)
 
 ui.run(title='RoSys')
